--- lib/translate.py
@@ -1,25 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-
-class TranslateBase:
-    """Translation base representation."""
-
-    start: float = -1.0
-    end: float = 1.0
-    duration_ms: int = 1000
-
-    def __call__(self, time_ms: int) -> float:
-        return self.ease(time_ms)
-
-    def function(self, t: float) -> int:
-        """User implemented translation function."""
-
-        return NotImplementedError
-
-    @classmethod
-    def ease(self, time_ms: int) -> float:
-        """Ease postion from current time in milliseconds."""
-
-        a = self.function(time_ms / self.duration_ms)
-        return a * (self.end - self.start) + self.start
